+++ addons/purchase_requisition/__manifest__.py
@@ -11,12 +11,14 @@ Manage calls for tenders and blanket orders. Calls for tenders are used to get
 competing offers from different vendors and select the best ones. Blanket orders
 are agreements you have with vendors to benefit from a predetermined pricing.
 """,
-    'depends' : ['purchase_stock'],
+    'depends': ['purchase_stock'],
     'demo': ['data/purchase_requisition_demo.xml'],
     'data': [
-        'security/purchase_tender.xml',
+        'security/purchase_requisition_security.xml',
         'security/ir.model.access.csv',
         'data/purchase_requisition_data.xml',
+        'views/product_views.xml',
+        'views/purchase_views.xml',
         'views/purchase_requisition_views.xml',
         'views/res_config_settings_views.xml',
         'report/purchase_requisition_report.xml',
