+++ addons/purchase_requisition/models/__init__.py
@@ -1,3 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import purchase
+from . import product
 from . import purchase_requisition
+from . import stock
